Include current line when indenting
Indentations have started from the next line to allow ease of highlighting by selecting from the end of the previous line, but this behavior may not be intuitive to all. This update includes the current line during indentation except when the caret is at a newline, allowing backward compatibility with the previous behavior.

- Indent current line unless current character is a newline, in which case indentation will start with the following line
- Change StringBuffer to StringBuilder during indentation for slight theoretical performance improvement
- Small formatting update for reverseIndexOf