Fix scrolling to a bookmark near the top of the document
Rather than scrolling the bookmark to the top of the viewport, simply scroll the view to the top of the document if the caret cannot be positioned in the middle of the viewport because the caret is too far toward the top of the document. Also clean up many comments.